+++ src/main/java/in/koreatech/koin/repository/shop/MenuCategoryMapRepository.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.repository;
+package in.koreatech.koin.repository.shop;
 
 import org.springframework.data.repository.Repository;
