+++ src/test/java/in/koreatech/koin/acceptance/ShopApiTest.java
@@ -11,9 +11,9 @@ import in.koreatech.koin.domain.shop.MenuCategory;
 import in.koreatech.koin.domain.shop.MenuCategoryMap;
 import in.koreatech.koin.domain.shop.MenuImage;
 import in.koreatech.koin.domain.shop.MenuOption;
-import in.koreatech.koin.repository.MenuCategoryMapRepository;
-import in.koreatech.koin.repository.MenuCategoryRepository;
-import in.koreatech.koin.repository.MenuRepository;
+import in.koreatech.koin.repository.shop.MenuCategoryMapRepository;
+import in.koreatech.koin.repository.shop.MenuCategoryRepository;
+import in.koreatech.koin.repository.shop.MenuRepository;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
